chore: clippy

+++ src/shared/src/day.rs
@@ -7,7 +7,7 @@ fn to_date(timestamp: &u64) -> OffsetDateTime {
     let nanos_remainder = nanoseconds % 1_000_000_000;
 
     OffsetDateTime::from_unix_timestamp(seconds).unwrap()
-        + Duration::nanoseconds(nanos_remainder as i64)
+        + Duration::nanoseconds(nanos_remainder)
 }
 
 pub fn day_of_the_year(timestamp: &u64) -> usize {
